backend/requirements/production.txt
Django==2.1.5
django-braces==1.13.0
django-extensions==2.1.4
django-tinymce==2.8.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
gunicorn==19.9.0
isort==4.3.4
olefile==0.46
Pillow==5.4.1
psycopg2==2.7.6.1
pytz==2018.9
six==1.12.0
